Updated relative scope modifier grammar
